move testing to the appropriate place
