+++ crates/goose/src/agents/agent.rs
@@ -196,7 +196,7 @@ impl Agent {
 
     /// Dispatch a single tool call to the appropriate client
     #[instrument(skip(self, tool_call, request_id), fields(input, output))]
-    pub(super) async fn dispatch_tool_call(
+    pub async fn dispatch_tool_call(
         &self,
         tool_call: mcp_core::tool::ToolCall,
         request_id: String,
@@ -400,7 +400,7 @@ impl Agent {
         (request_id, result)
     }
 
-    pub(crate) async fn handle_schedule_management(
+    pub async fn handle_schedule_management(
         &self,
         arguments: serde_json::Value,
         _request_id: String,
@@ -1461,211 +1461,5 @@ impl Agent {
             .expect("valid recipe");
 
         Ok(recipe)
-    }
-}
-
-#[cfg(test)]
-mod schedule_tool_tests {
-    use super::*;
-    use crate::scheduler::{ScheduledJob, SchedulerError};
-    use crate::scheduler_trait::SchedulerTrait;
-    use crate::session::storage::SessionMetadata;
-    use async_trait::async_trait;
-    use chrono::{DateTime, Utc};
-    use mcp_core::{Content, ToolError};
-    use serde_json::json;
-    use std::sync::Arc;
-
-    // Mock scheduler for testing
-    struct MockScheduler {
-        jobs: tokio::sync::Mutex<Vec<ScheduledJob>>,
-    }
-
-    impl MockScheduler {
-        fn new() -> Self {
-            Self {
-                jobs: tokio::sync::Mutex::new(Vec::new()),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl SchedulerTrait for MockScheduler {
-        async fn add_scheduled_job(&self, job: ScheduledJob) -> Result<(), SchedulerError> {
-            let mut jobs = self.jobs.lock().await;
-            jobs.push(job);
-            Ok(())
-        }
-
-        async fn list_scheduled_jobs(&self) -> Result<Vec<ScheduledJob>, SchedulerError> {
-            let jobs = self.jobs.lock().await;
-            Ok(jobs.clone())
-        }
-
-        async fn remove_scheduled_job(&self, id: &str) -> Result<(), SchedulerError> {
-            let mut jobs = self.jobs.lock().await;
-            if let Some(pos) = jobs.iter().position(|job| job.id == id) {
-                jobs.remove(pos);
-                Ok(())
-            } else {
-                Err(SchedulerError::JobNotFound(id.to_string()))
-            }
-        }
-
-        async fn pause_schedule(&self, _id: &str) -> Result<(), SchedulerError> {
-            Ok(())
-        }
-
-        async fn unpause_schedule(&self, _id: &str) -> Result<(), SchedulerError> {
-            Ok(())
-        }
-
-        async fn run_now(&self, _id: &str) -> Result<String, SchedulerError> {
-            Ok("test_session_123".to_string())
-        }
-
-        async fn sessions(
-            &self,
-            _sched_id: &str,
-            _limit: usize,
-        ) -> Result<Vec<(String, SessionMetadata)>, SchedulerError> {
-            Ok(vec![])
-        }
-
-        async fn update_schedule(
-            &self,
-            _sched_id: &str,
-            _new_cron: String,
-        ) -> Result<(), SchedulerError> {
-            Ok(())
-        }
-
-        async fn kill_running_job(&self, _sched_id: &str) -> Result<(), SchedulerError> {
-            Ok(())
-        }
-
-        async fn get_running_job_info(
-            &self,
-            _sched_id: &str,
-        ) -> Result<Option<(String, DateTime<Utc>)>, SchedulerError> {
-            Ok(None)
-        }
-    }
-
-    #[tokio::test]
-    async fn test_schedule_management_tool_list() {
-        let agent = Agent::new();
-        let mock_scheduler = Arc::new(MockScheduler::new());
-        agent.set_scheduler(mock_scheduler.clone()).await;
-
-        // Test list action
-        let arguments = json!({
-            "action": "list"
-        });
-
-        let result = agent
-            .handle_schedule_management(arguments, "test_req".to_string())
-            .await;
-        assert!(result.is_ok());
-
-        let content = result.unwrap();
-        assert_eq!(content.len(), 1);
-        if let Content::Text(text_content) = &content[0] {
-            assert!(text_content.text.contains("Scheduled Jobs:"));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_schedule_management_tool_no_scheduler() {
-        let agent = Agent::new();
-        // Don't set scheduler
-
-        let arguments = json!({
-            "action": "list"
-        });
-
-        let result = agent
-            .handle_schedule_management(arguments, "test_req".to_string())
-            .await;
-        assert!(result.is_err());
-
-        if let Err(ToolError::ExecutionError(msg)) = result {
-            assert!(msg.contains("Scheduler not available"));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_schedule_management_tool_run_now() {
-        let agent = Agent::new();
-        let mock_scheduler = Arc::new(MockScheduler::new());
-        agent.set_scheduler(mock_scheduler.clone()).await;
-
-        let arguments = json!({
-            "action": "run_now",
-            "job_id": "test_job"
-        });
-
-        let result = agent
-            .handle_schedule_management(arguments, "test_req".to_string())
-            .await;
-        assert!(result.is_ok());
-
-        let content = result.unwrap();
-        assert_eq!(content.len(), 1);
-        if let Content::Text(text_content) = &content[0] {
-            assert!(text_content
-                .text
-                .contains("Successfully started job 'test_job'"));
-            assert!(text_content.text.contains("test_session_123"));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_schedule_management_tool_dispatch() {
-        let agent = Agent::new();
-        let mock_scheduler = Arc::new(MockScheduler::new());
-        agent.set_scheduler(mock_scheduler.clone()).await;
-
-        // Test that the tool is properly dispatched through dispatch_tool_call
-        let tool_call = mcp_core::tool::ToolCall {
-            name: PLATFORM_MANAGE_SCHEDULE_TOOL_NAME.to_string(),
-            arguments: json!({
-                "action": "list"
-            }),
-        };
-
-        let (request_id, result) = agent
-            .dispatch_tool_call(tool_call, "test_dispatch".to_string())
-            .await;
-        assert_eq!(request_id, "test_dispatch");
-        assert!(result.is_ok());
-
-        let tool_result = result.unwrap();
-        // The result should be a future that resolves to the tool output
-        let output = tool_result.result.await;
-        assert!(output.is_ok());
-
-        let content = output.unwrap();
-        assert_eq!(content.len(), 1);
-        if let Content::Text(text_content) = &content[0] {
-            assert!(text_content.text.contains("Scheduled Jobs:"));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_schedule_management_tool_in_list_tools() {
-        let agent = Agent::new();
-        let tools = agent.list_tools(None).await;
-
-        // Check that the schedule management tool is included in the list
-        let schedule_tool = tools
-            .iter()
-            .find(|tool| tool.name == PLATFORM_MANAGE_SCHEDULE_TOOL_NAME);
-        assert!(schedule_tool.is_some());
-
-        let tool = schedule_tool.unwrap();
-        assert!(tool
-            .description
-            .contains("Manage scheduled recipe execution"));
     }
 }
