+++ crates/goose/src/agents/mod.rs
@@ -4,14 +4,10 @@ pub mod extension;
 pub mod extension_manager;
 mod large_response_handler;
 pub mod platform_tools;
-#[cfg(test)]
-mod private_tests;
 pub mod prompt_manager;
 mod reply_parts;
 mod router_tool_selector;
 mod router_tools;
-#[cfg(test)]
-mod test_support;
 
 mod tool_execution;
 mod tool_router_index_manager;
